feat: Add image analysis function using Firebase Functions and Google Generative AI
- Implement image upload and processing with Busboy and Sharp
- Add Google Generative AI integration for image analysis
- Configure CORS to handle cross-origin requests
- Handle POST requests to analyze images and provide a score with a detailed reason
- Include error handling and logging for debugging purposes
- Ensure temporary files are cleaned up after processing
