+++ functions/.gitignore
@@ -1,3 +1,5 @@
 node_modules/
 *.local
+firebase-debug.log
+ui-debug.log
 .env
