+++ functions/package.json
@@ -18,7 +18,10 @@
     "busboy": "^1.6.0",
     "cors": "^2.8.5",
     "firebase-admin": "^12.1.0",
-    "firebase-functions": "^5.0.0",
+    "firebase-functions": "^5.0.1",
+    "fs": "^0.0.1-security",
+    "os": "^0.1.2",
+    "path": "^0.12.7",
     "sharp": "^0.33.4"
   },
   "devDependencies": {
